iter14 tmp

+++ internal/repository/pgsqlrepo.go
@@ -29,16 +29,16 @@ func (pg *PostgresRepo) Add(ctx context.Context, user *models.User, name string)
 	hash := urlhasher.GetHash(name)
 
 	_, err := pg.Conn.Exec(ctx,
-		"INSERT INTO short_url (short_key, original_url) VALUES (@shortKey, @originalURL)",
-		pgx.NamedArgs{"shortKey": hash, "originalURL": name},
+		"INSERT INTO short_url (user_id, short_key, original_url) VALUES (@userId, @shortKey, @originalURL)",
+		pgx.NamedArgs{"userId": user.ID, "shortKey": hash, "originalURL": name},
 	)
 
 	if err != nil {
 		var pgErr *pgconn.PgError
 		if errors.As(err, &pgErr) && pgErr.Code == pgerrcode.UniqueViolation {
 			row := pg.Conn.QueryRow(ctx,
-				"SELECT short_key FROM short_url WHERE original_url=@originalUrl",
-				pgx.NamedArgs{"originalUrl": name},
+				"SELECT short_key FROM short_url WHERE user_id=@userId AND original_url=@originalUrl",
+				pgx.NamedArgs{"userId": user.ID, "originalUrl": name},
 			)
 
 			var shortKey string
@@ -66,13 +66,13 @@ func (pg *PostgresRepo) AddBatch(ctx context.Context, user *models.User, list *[
 
 	var entries [][]any
 	for _, v := range *list {
-		entries = append(entries, []any{v.ShortURL, v.OriginalURL})
+		entries = append(entries, []any{user.ID, v.ShortURL, v.OriginalURL})
 	}
 
 	_, err = pg.Conn.CopyFrom(
 		ctx,
 		pgx.Identifier{`short_url`},
-		[]string{"short_key", "original_url"},
+		[]string{"user_id", "short_key", "original_url"},
 		pgx.CopyFromRows(entries),
 	)
 
@@ -93,8 +93,8 @@ func (pg *PostgresRepo) GetByShortName(ctx context.Context, user *models.User, n
 	var originalURL string
 
 	row := pg.Conn.QueryRow(ctx,
-		"SELECT original_url FROM short_url WHERE short_key=@shortKey",
-		pgx.NamedArgs{"shortKey": name},
+		"SELECT original_url FROM short_url WHERE user_id=@userId AND short_key=@shortKey",
+		pgx.NamedArgs{"userId": user.ID, "shortKey": name},
 	)
 
 	err := row.Scan(&originalURL)
@@ -119,8 +119,8 @@ func (pg *PostgresRepo) IsReady(ctx context.Context) bool {
 func (pg *PostgresRepo) RemoveByOriginalURL(ctx context.Context, user *models.User, originalURL string) error {
 
 	_, err := pg.Conn.Exec(ctx,
-		"DELETE FROM short_url WHERE original_url=@original_url",
-		pgx.NamedArgs{"original_url": originalURL},
+		"DELETE FROM short_url WHERE user_id=@userId AND original_url=@original_url",
+		pgx.NamedArgs{"userId": user.ID, "original_url": originalURL},
 	)
 
 	return err
@@ -131,11 +131,12 @@ func createDBSchema(ctx context.Context, conn *pgxpool.Pool) error {
 	_, err := conn.Exec(ctx, `
 		CREATE TABLE IF NOT EXISTS short_url (
 		    id SERIAL NOT NULL PRIMARY KEY,
+		    user_id varchar(36) NOT NULL,
 		    short_key varchar(`+strconv.Itoa(urlhasher.HashLength)+`) UNIQUE NOT NULL,
 		    original_url text NOT NULL 
 		);
 
-		CREATE UNIQUE INDEX IF NOT EXISTS short_url_original_url_unique_idx ON short_url (original_url);
+		CREATE UNIQUE INDEX IF NOT EXISTS short_url_user_id_original_url_unique_idx ON short_url (user_id, original_url);
 	`)
 
 	if err != nil {
